@@ -7,19 +7,18 @@ homepage: https://github.com/OmkarDabade/markdown_editor_plus
 repository: https://github.com/OmkarDabade/markdown_editor_plus
 
 environment:
-  sdk: ">=3.1.5 <4.0.0"
+  sdk: ">=3.4.4 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   expandable: ^5.0.1
-  markdown: ^7.2.1
+  markdown: ^7.3.0
   flutter_markdown: ^0.6.19
-  font_awesome_flutter: ^10.7.0
+  font_awesome_flutter: ^10.10.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^4.0.0
 
-flutter: null
