@@ -1,7 +1,5 @@
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
-
 ///
 /// Constants defined for Emoji.
 ///
@@ -93,7 +91,7 @@ class Emoji {
   }
 
   @override
-  int get hashCode => hashValues(name.hashCode, code.hashCode);
+  int get hashCode => Object.hash(name.hashCode, code.hashCode);
 
   Emoji clone() {
     return Emoji(name, code);
